requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
pyarrow>=14.0.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
umap-learn>=0.5.5
plotly>=5.18.0
networkx>=3.2.1
python-dateutil>=2.8.2
